Don't optimise in logspace. Test to a separate file.
